+++ crates/miden-standards/build.rs
@@ -14,7 +14,6 @@ use miden_protocol_build_utils::{
     assemble_workspace,
     extract_all_masm_errors,
     generate_error_file,
-    generate_error_message_table,
 };
 
 // CONSTANTS
@@ -27,7 +26,6 @@ const ASM_COMPONENTS_DIR: &str = "components";
 
 const STANDARDS_ERRORS_RS_FILE: &str = "standards_errors.rs";
 const STANDARDS_ERRORS_ARRAY_NAME: &str = "STANDARDS_ERRORS";
-const MASM_ERROR_MESSAGES_RS_FILE: &str = "masm_error_messages.rs";
 
 // PRE-PROCESSING
 // ================================================================================================
@@ -119,7 +117,6 @@ fn generate_error_constants(asm_source_dir: &Path, build_dir: &str) -> Result<()
 
     let errors =
         extract_all_masm_errors(asm_source_dir).context("failed to extract all masm errors")?;
-    generate_error_message_table(Path::new(build_dir).join(MASM_ERROR_MESSAGES_RS_FILE), &errors)?;
     generate_error_file(
         ErrorModule {
             file_path: Path::new(build_dir).join(STANDARDS_ERRORS_RS_FILE),